fix: cast a int en el seed de demo (make_interval no acepta bigint)
`row_number()` devuelve bigint, así que `n * 9 - 5` llegaba a make_interval
como bigint y no encontraba overload: solo existe make_interval(days => int).
Falla al ejecutar, no al escribir — el seed no lo cubre ningún check.

+++ supabase/seed/demo-spread-dates.sql
@@ -21,7 +21,8 @@
 -- que no cambia, así que repartir dos veces da el mismo resultado.
 
 with ordenadas as (
-  select id, row_number() over (order by id) as n
+  -- `::int` porque row_number() es bigint y make_interval solo acepta int.
+  select id, (row_number() over (order by id))::int as n
     from public.bookings
    where status = 'completed'
      -- El tutor se identifica por una reserva suya conocida, no por nombre.
